+++ cuiman/src/cuiman/app/__init__.py
@@ -1,0 +1,25 @@
+#  Copyright (c) 2026 by the Eozilla team and contributors
+#  Permissions are hereby granted under the terms of the Apache 2.0 License:
+#  https://opensource.org/license/apache-2-0.
+
+from .display import (
+    AppColorScheme,
+    AppColorSchemeInput,
+    SerializedAppConfig,
+    ServiceProviderMeta,
+    ServiceProviderOption,
+    ServiceProviderType,
+    display_app_iframe,
+    display_packaged_app_iframe,
+)
+
+__all__ = [
+    "AppColorScheme",
+    "AppColorSchemeInput",
+    "SerializedAppConfig",
+    "ServiceProviderMeta",
+    "ServiceProviderOption",
+    "ServiceProviderType",
+    "display_app_iframe",
+    "display_packaged_app_iframe",
+]
